- Adicionada a classe Meu Local, e salvar meu local funcionando (falta mensagem modal de retorno)

+++ infomore-core/src/br/com/infomore/core/impl/dao/CategoriaDAO.java
@@ -8,7 +8,7 @@ import javax.persistence.Query;
 import br.com.infomore.dominio.Categoria;
 import br.com.infomore.dominio.EntidadeDominio;
 
-public class CategoriaDAO extends AbstractDAO<Long, Categoria> {
+public class CategoriaDAO extends AbstractDAO<Integer, Categoria> {
 
     public CategoriaDAO() {
 	super(Categoria.class);
